+++ api-lib/constants.js
@@ -7,9 +7,12 @@ export const ValidateProps = {
     bio: { type: 'string', minLength: 0, maxLength: 160 },
   },
   post: {
-    content: { type: 'string', minLength: 1, maxLength: 280 },
+    id: { type: 'string', minLength: 1, maxLength: 280 },
+    title: { type: 'string', minLength: 1, maxLength: 280 },
+    content: { type: 'string', minLength: 1, maxLength: 1000 },
+    published: { type: 'boolean' },
   },
   comment: {
-    content: { type: 'string', minLength: 1, maxLength: 280 },
+    content: { type: 'string', minLength: 1, maxLength: 1000 },
   },
 };