+++ components/Avatar/Avatar.jsx
@@ -1,8 +1,9 @@
 import styles from './Avatar.module.css';
 
-const Avatar = ({ size, username, url }) => {
+const Avatar = ({ style, size, username, url }) => {
   return (
     <img
+      style={style}
       className={styles.avatar}
       src={url || '/images/default_user.jpg'}
       alt={username}