+++ CHANGELOG.md
@@ -3,3 +3,6 @@
 ##### Release date: 3/5/2022
 - Init app version 2.0.0
 - improvement: update CLOUDINARY_URL and NODEMAILER_CONFIG
+
+##### Release date: 8/6/2022
+- new-feature: add function published and delete post